Began implementing the PCG plugin for Unreal
NewPCGgraph, and PCG volume are mere examples of how the plugin works: It takes a landscape the volume encompasses, and places the given node specified in the PCG blueprint at randomized points within that area. Given more mastery of the placement of those meshes, and the ability to have a list of meshes used instead of just one, we could definitely use this to generate the levels.
The example I created followed these instructions within the Documentation: https://dev.epicgames.com/documentation/en-us/unreal-engine/procedural-content-generation-framework-node-reference-in-unreal-engine
and also on this page are the Shape Grammar notes, which are almost definitely a good way to implement decisions on which rooms to spawn.

I also had to re-compile the C code, which is why it *looks* like I changed a lot more than I did. Consider this a 3 file change