Implemented New Design of Prefab Selection
This implementation uses a Stack, picking random prefabs until it either goes overbudget, or meets the sweet spot. If it's overbudget, it removes the selected prefab from the current selection pool, and rerolls. If it runs out of prefabs entirely in that selection, it removes the last selection, effectively going up one level higher. I hate recursion, which is why I used a stack instead