Working Constraint Satisfaction Algorithm
Previously it didn't actually do anything if the prefab pool was empty and the level grammar had something in it. This caused the method to just go to the end, add 31 to the level grammar and that would be the only thing that spawned. I added a condition right before 31 is added to the grammar that checks to see if the level grammar is empty and if the grammar stack has something in it. if so it adds the contents of the grammar stack to the level grammar. This check is for the case where the constraint could not be appropriately reached so the code go tit as close to reaching the constraint as possible.